refactor: replace internal draw API and update public types
- Replaced internal `_ikt{layout}_*` functions with `_inknit_draw_point32_{group_name}` and `_inknit_draw_point_with_clip32_{group_name}`.
  - Assembly implementations added for ARMv6-M, ARMv8-M, and x86 (IA-32); others fall back to C.
  - This is the first commit to introduce handwritten assembly (.s) files into the project.
  - New internal functions now use primitive types instead of struct arguments.
- Fixed export issues with pixel layouts: `"x1"`, `"x2"`, and `"x4"`.

BREAKING CHANGE: Public API argument types changed from `inknit_uint_t` to `int32_t`
- Renamed `inknit_{layout}_get` → `inknit_{layout}_get_pixel`
- Updated the following APIs to use `int32_t` for coordinates:
  - `inknit_{layout}_get_pixel`
  - `inknit_{layout}_draw_point`
  - `inknit_{layout}_draw_line`
  - `inknit_{layout}_draw_circle`

Other changes:
- Adjusted build options to allow `.s` files to compile correctly with Clang.
- Changed disassembly output for x86 and AMD64 targets to use Intel syntax in Release builds only.

+++ src/inknit/inc/inknit/types.h
@@ -108,7 +108,7 @@ struct inknit_image {
 
 	const uint16_t              stride;
 	const enum inknit_alignment alignment : 3;
-	const uint16_t              width     : 13;
+	const uint32_t              width     : 13;
 	const uint16_t              height;
 	void *INKNIT_RESTRICT       data;
 };
